# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Meridian Portfolio API",
    "version": "1.0.0",
    "description": "Interface for the Meridian front-end take-home. Meridian is a fictional platform where investors manage a portfolio of private-market (alternative) investments.\n\nThis is the only contract you need: build the front-end dashboard against these endpoints. Every request is served with a simulated latency (default 350ms) so loading and pending states are meaningful.\n\n- Collection endpoints return an envelope `{ data, meta }` (meta present when paginated) or `{ data }`.\n- Detail endpoints return the resource object directly.\n- Errors return `{ error, fields? }` — `fields` carries per-field validation messages on writes.\n\nUse **Try it out** on any endpoint below to see live responses."
  },
  "servers": [
    { "url": "http://localhost:4000/api", "description": "Local assessment API" }
  ],
  "tags": [
    { "name": "Portfolio", "description": "Aggregate portfolio metrics" },
    { "name": "Commitments", "description": "The investor's commitments to funds" },
    { "name": "Funds", "description": "Funds open for commitment" },
    { "name": "System", "description": "Service health" }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["System"],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "time": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/portfolio/summary": {
      "get": {
        "tags": ["Portfolio"],
        "summary": "Portfolio summary KPIs",
        "description": "Aggregate metrics computed across all commitments.",
        "responses": {
          "200": {
            "description": "Summary metrics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PortfolioSummary" }
              }
            }
          }
        }
      }
    },
    "/funds": {
      "get": {
        "tags": ["Funds"],
        "summary": "List funds",
        "description": "Combine filters freely; use `stage=open` for funds accepting commitments.",
        "parameters": [
          { "name": "strategy", "in": "query", "schema": { "$ref": "#/components/schemas/FundStrategy" } },
          { "name": "stage", "in": "query", "schema": { "$ref": "#/components/schemas/FundStage" } },
          { "name": "region", "in": "query", "schema": { "$ref": "#/components/schemas/FundRegion" } },
          { "name": "q", "in": "query", "description": "Free-text search over fund name and manager.", "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/sort" },
          { "$ref": "#/components/parameters/order" },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/pageSize" }
        ],
        "responses": {
          "200": {
            "description": "Paginated funds",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Paginated" },
                    { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Fund" } } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/funds/{id}/commitments": {
      "post": {
        "tags": ["Funds"],
        "summary": "Commit to a fund",
        "description": "Create a commitment to a fund. Enforces the fund's minimum investment and requires acknowledgement of terms. Field-level errors are returned under `fields`.",
        "parameters": [{ "$ref": "#/components/parameters/fundId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CommitRequest" },
              "example": { "amount": 500000, "acknowledgedTerms": true }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Commitment created (status `pending_close`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Commitment" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "Fund is closed to new commitments",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/commitments": {
      "get": {
        "tags": ["Commitments"],
        "summary": "List commitments",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/CommitmentStatus" } },
          { "name": "fundId", "in": "query", "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/sort" },
          { "$ref": "#/components/parameters/order" },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/pageSize" }
        ],
        "responses": {
          "200": {
            "description": "Paginated commitments",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Paginated" },
                    { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Commitment" } } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/commitments/{id}": {
      "get": {
        "tags": ["Commitments"],
        "summary": "Get a commitment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "cm_0001" }],
        "responses": {
          "200": {
            "description": "The commitment",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Commitment" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "sort": { "name": "sort", "in": "query", "description": "Field name to sort by.", "schema": { "type": "string" } },
      "order": { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "asc" } },
      "page": { "name": "page", "in": "query", "description": "1-based page number.", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "pageSize": { "name": "pageSize", "in": "query", "description": "Rows per page (max 100).", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 25 } },
      "fundId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "fund_northwind_iv" }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ValidationError": {
        "description": "Validation failed; see `fields` for per-field messages",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "error": "Please correct the highlighted fields.",
              "fields": { "amount": "Minimum commitment for this fund is $250,000.", "acknowledgedTerms": "You must acknowledge the fund terms to continue." }
            }
          }
        }
      }
    },
    "schemas": {
      "Paginated": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": {} },
          "meta": {
            "type": "object",
            "properties": {
              "page": { "type": "integer", "example": 1 },
              "pageSize": { "type": "integer", "example": 25 },
              "total": { "type": "integer", "example": 7 },
              "totalPages": { "type": "integer", "example": 1 }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "No fund with id \"fund_x\"" },
          "fields": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "description": "Present on validation failures; maps form field name to a human-readable message."
          }
        }
      },
      "FundStrategy": {
        "type": "string",
        "enum": ["Private Equity", "Venture Capital", "Growth Equity", "Private Credit", "Infrastructure", "Real Estate", "Secondaries"]
      },
      "FundStage": { "type": "string", "enum": ["open", "closing_soon", "closed"] },
      "FundRegion": { "type": "string", "enum": ["North America", "Europe", "Global"] },
      "CommitmentStatus": { "type": "string", "enum": ["active", "pending_close"] },
      "Fund": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "manager": { "type": "string" },
          "strategy": { "$ref": "#/components/schemas/FundStrategy" },
          "region": { "$ref": "#/components/schemas/FundRegion" },
          "vintage": { "type": "integer" },
          "stage": { "$ref": "#/components/schemas/FundStage" },
          "currency": { "type": "string", "example": "USD" },
          "fundSize": { "type": "integer" },
          "minInvestment": { "type": "integer" },
          "targetIrrPct": { "type": "number" },
          "managementFeePct": { "type": "number" },
          "carryPct": { "type": "number" },
          "investmentPeriodYears": { "type": "integer" },
          "fundTermYears": { "type": "integer" },
          "closeDate": { "type": "string", "format": "date" },
          "summary": { "type": "string" },
          "highlights": { "type": "array", "items": { "type": "string" } }
        },
        "example": {
          "id": "fund_northwind_iv",
          "name": "Northwind Buyout Fund IV",
          "manager": "Northwind Capital Partners",
          "strategy": "Private Equity",
          "region": "North America",
          "vintage": 2025,
          "stage": "open",
          "currency": "USD",
          "fundSize": 850000000,
          "minInvestment": 250000,
          "targetIrrPct": 18,
          "managementFeePct": 2,
          "carryPct": 20,
          "investmentPeriodYears": 5,
          "fundTermYears": 10,
          "closeDate": "2026-12-15",
          "summary": "Control-oriented buyouts of profitable middle-market industrials and business-services companies across North America.",
          "highlights": ["Targets EBITDA of $15M–$75M with proven operational levers"]
        }
      },
      "Commitment": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "fundId": { "type": "string" },
          "fundName": { "type": "string" },
          "manager": { "type": "string" },
          "strategy": { "$ref": "#/components/schemas/FundStrategy" },
          "committed": { "type": "number" },
          "called": { "type": "number" },
          "distributed": { "type": "number" },
          "nav": { "type": "number", "nullable": true, "description": "Net asset value. `null` until the first valuation is struck (e.g. a commitment that hasn't been called yet)." },
          "dateCommitted": { "type": "string", "format": "date" },
          "status": { "$ref": "#/components/schemas/CommitmentStatus" },
          "netIrrPct": { "type": "number", "nullable": true, "description": "`null` until there is enough history to compute a return." },
          "netMoic": { "type": "number", "nullable": true, "description": "`null` until there is enough history to compute a return." },
          "netDpi": { "type": "number", "nullable": true, "description": "`null` until there is enough history to compute a return." }
        },
        "example": {
          "id": "cm_0001",
          "fundId": "fund_silverpeak_v",
          "fundName": "Silverpeak Growth Equity V",
          "manager": "Silverpeak Growth",
          "strategy": "Growth Equity",
          "committed": 2000000,
          "called": 1400000,
          "distributed": 520000,
          "nav": 1680000,
          "dateCommitted": "2022-09-12",
          "status": "active",
          "netIrrPct": 19.4,
          "netMoic": 1.57,
          "netDpi": 0.37
        }
      },
      "PortfolioSummary": {
        "type": "object",
        "properties": {
          "totalCommitted": { "type": "number" },
          "totalCalled": { "type": "number" },
          "totalDistributed": { "type": "number" },
          "totalNav": { "type": "number" },
          "unfundedCommitment": { "type": "number" },
          "netMoic": { "type": "number" },
          "netDpi": { "type": "number" },
          "activeCommitments": { "type": "integer" },
          "currency": { "type": "string", "example": "USD" }
        },
        "example": {
          "totalCommitted": 10500000,
          "totalCalled": 5925000,
          "totalDistributed": 1620000,
          "totalNav": 5728000,
          "unfundedCommitment": 4575000,
          "netMoic": 1.24,
          "netDpi": 0.27,
          "activeCommitments": 5,
          "currency": "USD"
        }
      },
      "CommitRequest": {
        "type": "object",
        "required": ["amount", "acknowledgedTerms"],
        "properties": {
          "amount": { "type": "number", "description": "Commitment amount; must be ≥ the fund's minInvestment." },
          "acknowledgedTerms": { "type": "boolean", "description": "Must be true." }
        }
      }
    }
  }
}